Remove unneeded cast.

+++ tankgame/Tank.cpp
@@ -376,7 +376,7 @@ void Tank::Fire(float dTpressed)
     
     if(charge>=fireCost && fireTimer>fireRate)
     {
-        float ratio = (double)(z-TankHandler::GetSingleton().players[0].z)/(double)(x-TankHandler::GetSingleton().players[0].x);
+        float ratio = (z-TankHandler::GetSingleton().players[0].z)/(x-TankHandler::GetSingleton().players[0].x);
         float ryp=toDegrees(atan(ratio));
         
         if(TankHandler::GetSingleton().players[0].x<x)
